Create LaTeXCodeWholeNumbersChatGPTColourandTerms
Code created for LaTeX in mathematical format and for conversion to PDF, with space-separated thousands (following ISO/IEC 80000-1:2009) for clarity and uniformity in numerical spelling. To compile this LaTeX code, you will need a LaTeX distribution (such as TeX Live or MiKTeX) and a LaTeX editor (such as TeXstudio or Overleaf). Save the code as a .tex file, compile it.

NOTE: this last configuration did not work for running in Overleaf. Please leave your suggestion. Thanks